Add files via upload
This project focuses on creating a system that utilizes machine learning algorithms with Sony Spresense technology to efficiently harvest RF energy from the strongest energy sources. An LSTM model is implemented to analyze the frequency spectrum and pinpoint the areas with the highest energy density, ensuring that energy is captured from the most optimal sources.